+++ lib/rate_my_app.dart
@@ -1,4 +1,4 @@
-/// BubbleShowcase, created by Skyost
+/// RateMyApp, created by Skyost
 /// Github : https://github.com/Skyost/rate_my_app
 
 export 'package:rate_my_app/src/core.dart';